[refactor] Move QuickSearchModal
