+++ front-end/src/components/layout/QuickSearchModal.vue
@@ -34,8 +34,8 @@
 
 <script setup lang="ts">
 import { ref } from 'vue';
-import ModelCardDataTable from './content/ModelCardDataTable.vue';
-import { Pagination, SearchFilter } from './models';
+import ModelCardDataTable from 'src/components/content/ModelCardDataTable.vue';
+import { Pagination, SearchFilter } from '../models';
 
 const search = ref(false);
 const filter: SearchFilter = {};
