+++ front-end/src/components/layout/TheNavbar.vue
@@ -56,7 +56,7 @@ import { useAuthStore } from 'src/stores/auth-store';
 import { computed, ref } from 'vue';
 import { useRoute } from 'vue-router';
 
-import QuickSearchModal from 'src/components/QuickSearchModal.vue';
+import QuickSearchModal from 'src/components/layout/QuickSearchModal.vue';
 // import NotificationsMenu from 'src/components/NotificationsMenu.vue';
 import RouteBreadcrumbs from 'src/components/layout/RouteBreadcrumbs.vue';
 import DarkModeToggle from './DarkModeToggle.vue';
